Downgrade loglevel
If the nissan servers force us to log in again, that's not an error because 99% of the time we will try again and succeed. It's only an error if it subsequently fails.

This "ERROR" is the only error in my home assistant log. Downgrading this will silence the warning in normal operation.